fix: 메인 페이지 구조 정리
