+++ src/views/OfferInfoPage/constants/CHECK_OFFER_DATA.ts
@@ -1,0 +1,16 @@
+export const CHECK_OFFER_DATA = {
+  code: 200,
+  message: '제안서 승낙 성공',
+  data: {
+    applicationImgUrl: 'http://지원내역 캡쳐한거 클라가 넘겨준거 저장해둔거 보내쥼',
+    kakaoUrl: 'https://open.kakao.com/o/gukiuI2f',
+    designerInfo: {
+      imgUrl:
+        'https://mblogthumb-phinf.pstatic.net/20151022_73/rucifer25_1445501490504O4fSt_JPEG/%C1%B6%B4%CF%B5%AA%2C%C6%C0%B9%F6%C6%B0_%BF%B5%C8%AD%2C%B0%A1%C0%A7%BC%D5%2C%B8%ED%C0%DB_%BF%B5%C8%AD_%C3%DF%C3%B5_%282%29.jpg?type=w420',
+      shopName: '모디 헤어 강남점',
+      name: '너무 졸려 디자이너',
+      introduction:
+        '안녕하세요 모디 헤어 강남점 너무 졸려 디자이너 입니다람쥐*^^* 정말 너무 졸리고요, 너무 졸리네요. 다크서클이 언제 이렇게 내려온건지, 엄마가 아침에 절 보고 눈이 꺼지다 못해 뒤통수랑 하이파이브하겠다네요. 시커먼 다크서클은 턱끝까지 내려오고, 이러다가 요절하는 건 아닐지 걱정되는 요즘, 벌써 2024년이 돼서 한 살을 더 먹네요. 인생이 왜이리 빠른 건지 이대로 가다간 곧 가겠어요. 참 재밌네요나쁘지 않아, 행복해. 나는 괜찮다니까? 아무튼 일단 와보세요. 본업은잘 하는 편? 일단 믿어는 봐 해달라는 대로 드릴게^^',
+    },
+  },
+};
